Enum ra_hir_ty::Ty

pub enum Ty {
    Apply(ApplicationTy),
    Projection(ProjectionTy),
    Placeholder(TypeParamId),
    Bound(BoundVar),
    Infer(InferTy),
    Dyn(Arc<[GenericPredicate]>),
    Opaque(Arc<[GenericPredicate]>),
    Unknown,
}

A type.

See also the TyKind enum in rustc (librustc/ty/sty.rs), which represents the same thing (but in a different way).

This should be cheap to clone.

Variants

Apply(ApplicationTy)

A nominal type with (maybe 0) type parameters. This might be a primitive type like bool, a struct, tuple, function pointer, reference or several other things.

Projection(ProjectionTy)

A "projection" type corresponds to an (unnormalized) projection like <P0 as Trait<P1..Pn>>::Foo. Note that the trait and all its parameters are fully known.

Placeholder(TypeParamId)

A placeholder for a type parameter; for example, T in fn f<T>(x: T) {} when we're type-checking the body of that function. In this situation, we know this stands for some type, but don't know the exact type.

Bound(BoundVar)

A bound type variable. This is used in various places: when representing some polymorphic type like the type of function fn f<T>, the type parameters get turned into variables; during trait resolution, inference variables get turned into bound variables and back; and in Dyn the Self type is represented with a bound variable as well.

Infer(InferTy)

A type variable used during type checking.

Dyn(Arc<[GenericPredicate]>)

A trait object (dyn Trait or bare Trait in pre-2018 Rust).

The predicates are quantified over the Self type, i.e. Ty::Bound(0) represents the Self type inside the bounds. This is currently implicit; Chalk has the Binders struct to make it explicit, but it didn't seem worth the overhead yet.

Opaque(Arc<[GenericPredicate]>)

An opaque type (impl Trait).

The predicates are quantified over the Self type; see Ty::Dyn for more.

Unknown

A placeholder for a type which could not be computed; this is propagated to avoid useless error messages. Doubles as a placeholder where type variables are inserted before type checking, since we want to try to infer a better type here anyway -- for the IDE use case, we want to try to infer as much as possible even in the presence of type errors.

Methods

impl Ty

pub fn def_crates(
&self,
db: &dyn HirDatabase,
cur_crate: CrateId
) -> Option<ArrayVec<[CrateId; 2]>>

impl Ty

pub fn from_hir(ctx: &TyLoweringContext, type_ref: &TypeRef) -> Self

pub fn from_hir_ext(
ctx: &TyLoweringContext,
type_ref: &TypeRef
) -> (Self, Option<TypeNs>)

impl Ty

pub fn simple(ctor: TypeCtor) -> Ty

pub fn apply_one(ctor: TypeCtor, param: Ty) -> Ty

pub fn apply(ctor: TypeCtor, parameters: Substs) -> Ty

pub fn unit() -> Self

pub fn as_reference(&self) -> Option<(&Ty, Mutability)>

pub fn strip_references(&self) -> &Ty

pub fn as_adt(&self) -> Option<(AdtId, &Substs)>

pub fn as_tuple(&self) -> Option<&Substs>

pub fn as_callable(&self) -> Option<(CallableDef, &Substs)>

pub fn dyn_trait_ref(&self) -> Option<&TraitRef>

If this is a dyn Trait type, this returns the Trait part.

pub fn apply_substs(self, substs: Substs) -> Ty

If this is a type with type parameters (an ADT or function), replaces the Substs for these type parameters with the given ones. (So e.g. if self is Option<_> and the substs contain u32, we'll have Option<u32> afterwards.)

pub fn substs(&self) -> Option<Substs>

Returns the type parameters of this type if it has some (i.e. is an ADT or function); so if self is Option<u32>, this returns the u32.

pub fn inherent_trait(&self) -> Option<TraitId>

If this is an impl Trait or dyn Trait, returns that trait.